Fixed Errors

+++ src/main/java/web/project/track/repo/DailyLogRepository.java
@@ -2,9 +2,8 @@ package web.project.track.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import web.project.track.entity.DailyLog;
-
 import java.util.List;
 
-public interface DailyLogRepository extends JpaRepository<DailyLog, Long> {
-    List<DailyLog> findByStudentId(int studentId); // Custom query method
+public interface DailyLogRepository extends JpaRepository<DailyLog, Integer> {
+    List<DailyLog> findByStudentId(int studentId);
 }
